Spring boot
Built first Java application that can actually do something. Used spring boot to build a basic API that makes request to a Postgresql database that was setup with docker. You can get, post and delete customers that are within the database. Just a basic project to work on my Java skills