Library Catalog System
This is the bones of a library catalog system. With the system connected to a GUI, you can look up what books are within the system, look up the authors within the system, look up book collections within the library associated to a specific author and you can see the borrow history of a book. There is no GUI connected to this system. Everything that is in main.java is just book data being plugged in and testing of certain methods. Will probably not add a gui to it but the methods and functionality is there so it would just be a plug and play scenario.