SmallMart-BE
Developed a full stack program that consisted of a java backend and a reactjs front end. The emphasis of this project was on the Java backend so the front end is very minimal and is not pretty. Main purpose was to just have something to run against the back end so it is not pretty. User can sell products and buy products. It is not super in-depth but it does the basics and gave me a good introduction and to using Springboot and how to get and post json data using class objects.