Let's Catchup BE
This is the backend for a social media app i created. The purpose of the project was to further build my java skills so the front-end was built only to verify everything worked. The front-end was built in react and typescript. The backend uses Java, Springboot and MongoDB. The application allows a user to view their profile, see what they have posted as well as see what their friends have posted. They can accept friend request, comment on post and like post.