Stock Trading app
Trading app built in java and springboot. Can do the basics like look up stocks, add and remove from portfolio as well as view the portfolio. Some more things could be worked on like what specific day to see. But in general it is a solid API to work with if you want to add a gui to it.